requirements.txt
# Core dependencies
pyspark>=3.4.0
python-dotenv>=1.0.0

# MySQL connectivity
mysql-connector-python>=8.0.33

# Data processing and trends
pytrends>=4.9.0
pandas>=1.5.0

# Django and API dependencies
Django>=4.2.0
djangorestframework>=3.14.0
django-filter>=23.2
django-cors-headers>=4.0.0
mysqlclient>=2.2.0
drf-spectacular>=0.26.0
djangorestframework-simplejwt>=5.3.0

# HTTP requests for API calls
requests>=2.28.0

# Web scraping dependencies
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data normalization
pycountry>=24.6.0

# Optional: For enhanced logging and monitoring
py4j>=0.10.9

